requirements.txt
git+https://github.com/Rapptz/discord.py
git+https://github.com/Gorialis/jishaku
aiohttp==3.7.4
import_expression
discord-ext-menus
asyncpg